[YUNIKORN-592] test: move the testing function into the relevant test code file (#662)
Closes: #662
